feat: decoretors for control acess add

+++ src/users/users.controller.ts
@@ -6,11 +6,18 @@ import {
   Patch,
   Param,
   Delete,
+  UseGuards,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
-import { ApiTags } from '@nestjs/swagger';
+import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
+import { AuthGuard } from '@nestjs/passport';
+import { isAdmin } from 'src/auth/decorators/isAdmsin.decorator';
+import { Users } from './entities/user.entity';
+import { isTheUser } from 'src/auth/decorators/isTheUser.decorator';
+import { ParameterLocation, ParameterObject } from '@nestjs/swagger/dist/interfaces/open-api-spec.interface';
 @ApiTags('users')
 @Controller('users')
 export class UsersController {
@@ -22,22 +29,40 @@ export class UsersController {
   }
 
   @Get()
-  findAll() {
+  @UseGuards(AuthGuard())
+  @ApiBearerAuth()
+  findAll(@isAdmin() user:Users) {
     return this.usersService.findAll();
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.usersService.findOne(id);
+  @UseGuards(AuthGuard())
+  @ApiBearerAuth()
+ 
+  findOne(@isTheUser() user:Users, @Param('id') id: string) {
+    if(user.id!== id){ throw new UnauthorizedException(
+      'user not have permission to access this route',
+    );}
+  return this.usersService.findOne(id);
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+  @UseGuards(AuthGuard())
+  @ApiBearerAuth()
+  update(@isTheUser() user:Users,@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+    if(user.id!== id){ throw new UnauthorizedException(
+      'user not have permission to access this route',
+    );}
     return this.usersService.update(id, updateUserDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  @UseGuards(AuthGuard())
+  @ApiBearerAuth()
+  remove(@isTheUser() user:Users,@Param('id') id: string) {
+    if(user.id!== id){ throw new UnauthorizedException(
+      'user not have permission to access this route',
+    );}
     return this.usersService.remove(id);
   }
 }
